+++ mcp_tools/network_recon/enhanced_network_scanning.py
@@ -78,84 +78,6 @@ def register_enhanced_network_scanning_tools(mcp, hexstrike_client, logger):
         return result
 
     @mcp.tool()
-    def autorecon_comprehensive(target: str, output_dir: str = "/tmp/autorecon",
-                               port_scans: str = "top-100-ports", service_scans: str = "default",
-                               heartbeat: int = 60, timeout: int = 300,
-                               additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute AutoRecon for comprehensive automated reconnaissance.
-
-        Args:
-            target: The target IP address or hostname
-            output_dir: Output directory for results
-            port_scans: Port scan configuration
-            service_scans: Service scan configuration
-            heartbeat: Heartbeat interval in seconds
-            timeout: Timeout for individual scans
-            additional_args: Additional AutoRecon arguments
-
-        Returns:
-            Comprehensive automated reconnaissance results
-        """
-        data = {
-            "target": target,
-            "output_dir": output_dir,
-            "port_scans": port_scans,
-            "service_scans": service_scans,
-            "heartbeat": heartbeat,
-            "timeout": timeout,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔄 Starting AutoRecon: {target}")
-        result = hexstrike_client.safe_post("api/tools/autorecon", data)
-        if result.get("success"):
-            logger.info(f"✅ AutoRecon completed for {target}")
-        else:
-            logger.error(f"❌ AutoRecon failed for {target}")
-        return result
-
-    @mcp.tool()
-    def enum4linux_ng_advanced(target: str, username: str = "", password: str = "",
-                               domain: str = "", shares: bool = True, users: bool = True,
-                               groups: bool = True, policy: bool = True,
-                               additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Enum4linux-ng for advanced SMB enumeration with enhanced logging.
-
-        Args:
-            target: The target IP address
-            username: Username for authentication
-            password: Password for authentication
-            domain: Domain for authentication
-            shares: Enumerate shares
-            users: Enumerate users
-            groups: Enumerate groups
-            policy: Enumerate policies
-            additional_args: Additional Enum4linux-ng arguments
-
-        Returns:
-            Advanced SMB enumeration results
-        """
-        data = {
-            "target": target,
-            "username": username,
-            "password": password,
-            "domain": domain,
-            "shares": shares,
-            "users": users,
-            "groups": groups,
-            "policy": policy,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Enum4linux-ng: {target}")
-        result = hexstrike_client.safe_post("api/tools/enum4linux-ng", data)
-        if result.get("success"):
-            logger.info(f"✅ Enum4linux-ng completed for {target}")
-        else:
-            logger.error(f"❌ Enum4linux-ng failed for {target}")
-        return result
-
-    @mcp.tool()
     def rpcclient_enumeration(target: str, username: str = "", password: str = "",
                              domain: str = "", commands: str = "enumdomusers;enumdomgroups;querydominfo",
                              additional_args: str = "") -> Dict[str, Any]:
@@ -218,101 +140,8 @@ def register_enhanced_network_scanning_tools(mcp, hexstrike_client, logger):
             logger.error(f"❌ nbtscan failed for {target}")
         return result
 
-    @mcp.tool()
-    def arp_scan_discovery(target: str = "", interface: str = "", local_network: bool = False,
-                          timeout: int = 500, retry: int = 3, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute arp-scan for network discovery with enhanced logging.
 
-        Args:
-            target: The target IP range (if not using local_network)
-            interface: Network interface to use
-            local_network: Scan local network
-            timeout: Timeout in milliseconds
-            retry: Number of retries
-            additional_args: Additional arp-scan arguments
 
-        Returns:
-            Network discovery results via ARP scanning
-        """
-        data = {
-            "target": target,
-            "interface": interface,
-            "local_network": local_network,
-            "timeout": timeout,
-            "retry": retry,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting arp-scan: {target if target else 'local network'}")
-        result = hexstrike_client.safe_post("api/tools/arp-scan", data)
-        if result.get("success"):
-            logger.info(f"✅ arp-scan completed")
-        else:
-            logger.error(f"❌ arp-scan failed")
-        return result
 
-    @mcp.tool()
-    def responder_credential_harvest(interface: str = "eth0", analyze: bool = False,
-                                   wpad: bool = True, force_wpad_auth: bool = False,
-                                   fingerprint: bool = False, duration: int = 300,
-                                   additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Responder for credential harvesting with enhanced logging.
 
-        Args:
-            interface: Network interface to use
-            analyze: Analyze mode only
-            wpad: Enable WPAD rogue proxy
-            force_wpad_auth: Force WPAD authentication
-            fingerprint: Fingerprint mode
-            duration: Duration to run in seconds
-            additional_args: Additional Responder arguments
-
-        Returns:
-            Credential harvesting results
-        """
-        data = {
-            "interface": interface,
-            "analyze": analyze,
-            "wpad": wpad,
-            "force_wpad_auth": force_wpad_auth,
-            "fingerprint": fingerprint,
-            "duration": duration,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Responder on interface: {interface}")
-        result = hexstrike_client.safe_post("api/tools/responder", data)
-        if result.get("success"):
-            logger.info(f"✅ Responder completed")
-        else:
-            logger.error(f"❌ Responder failed")
-        return result
-
-    @mcp.tool()
-    def volatility_analyze(memory_file: str, plugin: str, profile: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Volatility for memory forensics analysis with enhanced logging.
-
-        Args:
-            memory_file: Path to memory dump file
-            plugin: Volatility plugin to use
-            profile: Memory profile to use
-            additional_args: Additional Volatility arguments
-
-        Returns:
-            Memory forensics analysis results
-        """
-        data = {
-            "memory_file": memory_file,
-            "plugin": plugin,
-            "profile": profile,
-            "additional_args": additional_args
-        }
-        logger.info(f"🧠 Starting Volatility analysis: {plugin}")
-        result = hexstrike_client.safe_post("api/tools/volatility", data)
-        if result.get("success"):
-            logger.info(f"✅ Volatility analysis completed")
-        else:
-            logger.error(f"❌ Volatility analysis failed")
-        return result
